am64x/am243x: sa2ul: fix processing of large buffers
The SA2UL driver processes the buffers, larger than or equal to the two
chunks of maximum packet size, incorrectly. This, for example, leads to
the incorrect SHA512 hash calculation for buffers larger than 8MB.

This is because of the following reasons:

- The driver does not reinitialize the `donedataLen` to 0 for each
  iteration of the for loop meant for processing the maximum sized
  packet.
- The driver advances the `ptrInput` and `ptrOutput` pointers
  incorrectly leading the engine to read garbage input.

Fix the issue by initializing the above variables correctly.

Additionally, fix the typos in the certain cache related comments.

Fixes: SITSW-8683, SITSW-8686

Signed-off-by: Prashant Shivhare <p-shivhare@ti.com>